Revert "#175 tests for adding change logs to enquiry"
