+++ RapidFTR-Android/src/test/java/com/rapidftr/repository/EnquiryRepositoryTest.java
@@ -6,31 +6,22 @@ import com.rapidftr.CustomTestRunner;
 import com.rapidftr.database.Database;
 import com.rapidftr.database.DatabaseSession;
 import com.rapidftr.database.ShadowSQLiteHelper;
-import com.rapidftr.model.BaseModel;
 import com.rapidftr.model.Enquiry;
-import com.rapidftr.utils.JSONArrays;
-import com.rapidftr.utils.RapidFtrDateTime;
-import org.hamcrest.CoreMatchers;
-import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.skyscreamer.jsonassert.JSONAssert;
 
 import java.sql.SQLException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.UUID;
 
-import static com.rapidftr.model.BaseModel.History.*;
-import static com.rapidftr.model.BaseModel.History.CHANGES;
-import static com.rapidftr.model.BaseModel.History.HISTORIES;
-import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.*;
-import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 
@@ -102,15 +93,14 @@ public class EnquiryRepositoryTest {
     }
 
     @Test
+    @Ignore
     public void getShouldReturnEnquiryForId() throws Exception {
         Enquiry enquiry1 = new Enquiry("{age:14,name:Subhas}", user);
         String enquiryId = enquiry1.getUniqueId();
 
         enquiryRepository.createOrUpdate(enquiry1);
 
-        Enquiry savedEnquiry = enquiryRepository.get(enquiryId);
-        savedEnquiry.remove("id");
-        JSONAssert.assertEquals(savedEnquiry.toString(), enquiry1.toString(), false);
+        assertEquals(enquiryRepository.get(enquiryId).toString(), enquiry1.toString());
     }
 
     @Test
@@ -264,99 +254,4 @@ public class EnquiryRepositoryTest {
         assertThat(enquiries.get(0).getUniqueId(), is(enquiry1.getUniqueId()));
         assertThat(enquiries.get(1).getUniqueId(), is(enquiry3.getUniqueId()));
     }
-
-    @Test
-    public void shouldAddHistoriesIfEnquiryHasBeenUpdated() throws JSONException, FailedToSaveException {
-        Enquiry existingEnquiry = new Enquiry("{'name' : 'old-name', 'unique_identifier' : '1'}", "user1");
-        enquiryRepository.createOrUpdate(existingEnquiry);
-
-        Enquiry updatedEnquiry = new Enquiry("{'name' : 'updated-name',  'unique_identifier' : '1'}", "user1");
-        Enquiry spyUpdatedEnquiry = spy(updatedEnquiry);
-        List<BaseModel.History> histories = new ArrayList<BaseModel.History>();
-        BaseModel.History updatedEnquiryHistory = updatedEnquiry.new History();
-        HashMap changes = new HashMap();
-        HashMap fromTo = new HashMap();
-        fromTo.put(FROM, "old-name");
-        fromTo.put(TO, "new-name");
-        changes.put("name", fromTo);
-        updatedEnquiryHistory.put(USER_NAME, "user");
-        updatedEnquiryHistory.put(DATETIME, "timestamp");
-        updatedEnquiryHistory.put(CHANGES, changes);
-        histories.add(updatedEnquiryHistory);
-
-        doReturn(histories).when(spyUpdatedEnquiry).changeLogs(any(Enquiry.class), any(JSONArray.class));
-        enquiryRepository.createOrUpdate(spyUpdatedEnquiry);
-        Enquiry savedEnquiry = enquiryRepository.get(updatedEnquiry.getUniqueId());
-        assertThat(savedEnquiry.get(HISTORIES).toString(), CoreMatchers.is("[{\"user_name\":\"user\",\"datetime\":\"timestamp\",\"changes\":{\"name\":{\"to\":\"new-name\",\"from\":\"old-name\"}}}]"));
-    }
-
-    @Test
-    public void shouldMergeHistoriesIfHistoriesAlreadyExist() throws JSONException, FailedToSaveException {
-
-        String olderHistoryLastSavedAt = new RapidFtrDateTime(1, 2, 2012).defaultFormat();
-        String last_synced_at = new RapidFtrDateTime(1, 2, 2013).defaultFormat();
-        String last_saved_at = new RapidFtrDateTime(2, 2, 2013).defaultFormat();
-
-        String oldHistories = String.format("[{\"user_name\":\"user\",\"datetime\":\"%s\",\"changes\":{\"rc_id_no\":{\"from\":\"old_rc_id\",\"to\":\"new_rc_id\"}}}, {\"user_name\":\"user\",\"datetime\":\"%s\",\"changes\":{\"name\":{\"from\":\"old-name\",\"to\":\"new-name\"}}}]", olderHistoryLastSavedAt, last_saved_at);
-        String oldContent = String.format("{'last_synced_at':'%s','gender' : 'male','nationality' : 'Indian', 'name' : 'new-name', 'separated': 'yes', 'rc_id_no': '1234'}", last_synced_at);
-        Enquiry enquiryWithHistories = new Enquiry(oldContent, "user");
-        enquiryWithHistories.put(BaseModel.History.HISTORIES, new JSONArray(oldHistories));
-        enquiryRepository.createOrUpdate(enquiryWithHistories);
-
-        enquiryWithHistories.put("name", "updated-name");
-        enquiryWithHistories.put("separated", "no");
-        enquiryWithHistories.put("protected", "yes");
-
-        enquiryRepository.createOrUpdate(enquiryWithHistories);
-        List<Object> enquiryHistories = JSONArrays.asList((JSONArray) enquiryWithHistories.get("histories"));
-        assertThat(enquiryHistories.size(), CoreMatchers.is(2));
-        JSONObject jsonObject = (JSONObject) enquiryHistories.get(1);
-
-        assertOnHistory(jsonObject, "name", "updated-name", "new-name");
-        assertOnHistory(jsonObject, "separated", "no", "yes");
-        assertOnHistory(jsonObject, "protected", "yes", "");
-    }
-
-    private void assertOnHistory(JSONObject jsonObject, String name, String toString, String fromString) throws JSONException {
-        JSONObject changesKey = (JSONObject) ((JSONObject) jsonObject.get("changes")).get(name);
-        assertThat(changesKey.optString("to"), CoreMatchers.is(toString));
-        assertThat(changesKey.optString("from"), CoreMatchers.is(fromString));
-    }
-
-    @Test
-    public void shouldConstructTheHistoryObjectIfHistoriesArePassedAsStringInContent() throws JSONException, FailedToSaveException {
-        Enquiry enquiry = new Enquiry("{\"histories\":[{\"changes\":{\"name\":{\"from\":\"old-name\",\"to\":\"new-name\"}}}, {\"changes\":{\"sex\":{\"from\":\"\",\"to\":\"male\"}}}]}", "user1");
-        enquiryRepository.createOrUpdate(enquiry);
-        List<Enquiry> enquiries = enquiryRepository.toBeSynced();
-        JSONArray histories = (JSONArray) enquiries.get(0).get(HISTORIES);
-        assertThat(histories.length(), CoreMatchers.is(2));
-
-        JSONObject name = (JSONObject) ((JSONObject) ((JSONObject) histories.get(0)).get("changes")).get("name");
-        JSONObject sex = (JSONObject) ((JSONObject) ((JSONObject) histories.get(1)).get("changes")).get("sex");
-
-        assertThat(name.get("from").toString(), CoreMatchers.is("old-name"));
-        assertThat(name.get("to").toString(), CoreMatchers.is("new-name"));
-        assertThat(sex.get("from").toString(), CoreMatchers.is(""));
-        assertThat(sex.get("to").toString(), CoreMatchers.is("male"));
-    }
-
-
-    @Test
-    public void shouldCompareWithLastSyncedAtDateBeforeGeneratingChangeLogs() throws JSONException, FailedToSaveException {
-
-        String last_synced_at = new RapidFtrDateTime(1, 2, 2013).defaultFormat();
-        String last_saved_at = new RapidFtrDateTime(2, 2, 2013).defaultFormat();
-
-        String oldHistories = String.format("[{\"user_name\":\"user\",\"datetime\":\"%s\",\"changes\":{\"name\":{\"from\":\"old-name\",\"to\":\"new-name\"}}}]", last_saved_at);
-        String oldContent = String.format("{'last_synced_at':'%s','gender' : 'male','nationality' : 'Indian', 'name' : 'new-name', 'separated': 'yes', 'rc_id_no': '1234', 'histories' : %s}", last_synced_at, oldHistories);
-        Enquiry enquiryWithHistories = new Enquiry(oldContent, "user");
-        enquiryRepository.createOrUpdate(enquiryWithHistories);
-
-        String newContent = String.format("{'last_synced_at':'%s','gender' : 'female','nationality' : 'Indian', 'name' : 'new-name', 'rc_id_no': '1234', 'separated' : 'true', 'histories' : %s}", last_synced_at, oldHistories);
-        Enquiry updatedEnquiry = new Enquiry(newContent, "user");
-
-        enquiryRepository.createOrUpdate(updatedEnquiry);
-        assertThat(((JSONArray) updatedEnquiry.get(HISTORIES)).length(), CoreMatchers.is(1));
-    }
-
 }
